% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_progression_bound.R
\name{binom_progression_bound}
\alias{binom_progression_bound}
\title{Bound for Binomial Progression}
\usage{
binom_progression_bound(start, count)
}
\arguments{
\item{start}{Non-negative real number, the starting term of the arithmetic progression in the denominator.}

\item{count}{The number of terms in the denominator.}
}
\value{
The infimum of all values of x for which the expression is well-defined, i.e. the denominator is non-zero.
}
\description{
Returns the bound for x allowed in the expression 1/(start * (start+x) * ... * (start+(count-1)*x)).
A small subfunction to make the calculations in the next function a bit easier.
The edge cases are as follows. If the count is equal to zero, all x are allowed, so the infimum is -Inf.
If the count is greater than zero, and the starting term is equal to zero, then all x are disallowed.
Then the infimum is Inf.
}
