% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_iQCC_for_selected_samples.R
\name{compute_iQCC_for_selected_samples}
\alias{compute_iQCC_for_selected_samples}
\title{Overdispersion measures (iQCCs)}
\usage{
compute_iQCC_for_selected_samples(df, reps, inf_Q = 1, sup_Q = 1)
}
\arguments{
\item{df}{Allele counts dataframe: with 2n+1 columns, "ID" and 2n columns with ref & alt counts (rep1_ref, rep1_alt, rep2_ref, rep2_alt, ...)}

\item{reps}{Optional (default=NA, all replicates), a vector of replicate numbers for which the analysis should be applied}

\item{inf_Q}{Optional (default=1) Only use the data rows for which inf_Q yields a well-defined distribution. The default inf_Q=1 simplifies to the binomial distribution, therefore it does not restrict the input.}

\item{sup_Q}{Optional (default=1) Only use the data rows for which sup_Q yields a well-defined distribution. The default sup_Q=1 does not restrict the input. The parameter can be used to effectively filter the rows with a low coverage, e.g. setting sup_Q=30 filters out the data rows with the coverage less than 30.}
}
\value{
A list of iQCC triplets for each sample within selected: lower, upper and exact estimate.
}
\description{
Calculates iQCC values for selected samples in allelic table.
}
