% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dextbetabinom.R
\name{dextbetabinom}
\alias{dextbetabinom}
\title{The Probability Density Function of the Extended Beta-Binomial Distribution}
\usage{
dextbetabinom(
  x,
  size,
  alpha = NA,
  beta = NA,
  d = NA,
  AI = NA,
  Q = NA,
  log.p = FALSE
)
}
\arguments{
\item{x}{The number of successful draws.}

\item{size}{The total number of draws.}

\item{alpha}{Optional (default=NA) The prior number of successful draws in Polya urn model.}

\item{beta}{Optional (default=NA) The prior number of unsuccessful draws in Polya urn model.}

\item{d}{Optional (default=NA) The prior number of unsuccessful draws in Polya urn model.}

\item{AI}{Optional (default=NA) The proportion between alpha and beta.}

\item{Q}{Optional (default=NA) The overdispersion parameter.}

\item{log.p}{Optional (default=FALSE) Whether to return the pdf or the logarithm of it. Get the unchanged value of the pdf by default.}
}
\value{
The value of the pdf of the extended beta-binomial distribution.
}
\description{
Returns the value of the pdf of the extended beta-binomial distribution. The function will accept the parameters either using the alpha, beta and d parameters, or using the AI and Q parameters.
}
