% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dextbetabinom_log_derivative_Q.R
\name{dextbetabinom_log_derivative_Q}
\alias{dextbetabinom_log_derivative_Q}
\title{The Log Derivative of the Extended Beta-Binomial Distribution}
\usage{
dextbetabinom_log_derivative_Q(x, size, AI, Q, force_calculation = FALSE)
}
\arguments{
\item{x}{The number of successful draws.}

\item{size}{The total number of draws.}

\item{AI}{The proportion between alpha and beta.}

\item{Q}{The overdispersion parameter.}

\item{force_calculation}{Optional (default=FALSE) Whether to skip the sanity checks for whether the parameters are in the definition range and calculate the derivative directly. Do the checks by default.}
}
\value{
The value of the logarithm of the pdf of the extended beta-binomial distribution.
}
\description{
Returns the log derivative of the pdf of the extended beta-binomial distribution with regards to Q.
}
