% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dextbetabinom_log_likelihood_derivative_Q.R
\name{dextbetabinom_log_likelihood_derivative_Q}
\alias{dextbetabinom_log_likelihood_derivative_Q}
\title{The Log Likelihood Derivative using the Extended Beta-Binomial Distribution}
\usage{
dextbetabinom_log_likelihood_derivative_Q(
  mat_counts,
  pat_counts,
  AIs,
  Q,
  force_calculation = FALSE,
  na.rm = F
)
}
\arguments{
\item{mat_counts}{A vector of the maternal counts.}

\item{pat_counts}{A vector of the paternal counts.}

\item{AIs}{A vector of true allelic imbalance proportions.}

\item{Q}{The overdispersion parameter.}

\item{force_calculation}{Optional (default=FALSE) Whether to skip the sanity checks for whether the parameters are in the definition range and calculate the derivative directly. Do the checks by default.}

\item{na.rm}{Optional (default=FALSE) Whether to remove the parameter rows with at least one NA. Do not remove by default.}
}
\value{
The value of the pdf of the extended beta-binomial distribution.
}
\description{
Returns the log likelihood derivative using the extended beta-binomial distribution for an array of parameters.
}
