% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dextbetabinom_main_ranges_d.R
\name{dextbetabinom_main_ranges_d}
\alias{dextbetabinom_main_ranges_d}
\title{Definition and Positivity Range of Extended Deta-Binomial Distribution In Terms of d}
\usage{
dextbetabinom_main_ranges_d(x, size, alpha, beta, AI = NA)
}
\arguments{
\item{x}{The number of successful draws.}

\item{size}{The total number of draws.}

\item{alpha}{The prior number of successful draws in Polya urn model.}

\item{beta}{The prior number of unsuccessful draws in Polya urn model.}

\item{AI}{Optional (default=NA) A parameter to indicate the proportion between alpha in beta in the edge cases like alpha = beta = 0.}
}
\value{
A matrix 2x3 with the basic information about the values of the probability density function as a function of d. The elements result(1,1) and result(1,2) denote the interval of the values of d where the pdf is defined. The elements result(2,1) and result(2,2) denote the interval of the values of d where the pdf is positive. The entry result(1,3) contains the value of the pdf if it does not depend on d.
}
\description{
Returns the intervals where the probability density function of the extended beta-binomial distribution is defined, and where it is positive, in terms of the intermediate variable d.
Here we strive to consider as many edge cases carefully as we can, e.g. if x = 0 and size = 0, we aknowledge that the probability value is 1 even if alpha and beta are NA.
}
