% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dextbetabinom_max_likelihood_Q.R
\name{dextbetabinom_max_likelihood_Q}
\alias{dextbetabinom_max_likelihood_Q}
\title{Best Fitting Q According to the Extended Beta-Binomial Model}
\usage{
dextbetabinom_max_likelihood_Q(allele_counts, AI_vect, inf_Q = 1, sup_Q = 1)
}
\arguments{
\item{allele_counts}{A dataframe with the maternal and paternal counts.}

\item{AI_vect}{A vector of true allelic imbalance proportions.}

\item{inf_Q}{Optional (default=1) Only use the data rows for which inf_Q yields a well-defined distribution. The default inf_Q=1 simplifies to the binomial distribution, therefore it does not restrict the input.}

\item{sup_Q}{Optional (default=1) Only use the data rows for which sup_Q yields a well-defined distribution. The default sup_Q=1 does not restrict the input. The parameter can be used to effectively filter the rows with a low coverage, e.g. setting sup_Q=30 filters out the data rows with the coverage less than 30.}
}
\value{
The fitted value of Q.
}
\description{
Finds a Q maximizing the likelihood using several attempts of gradient descend. One attempt may not succeed because of data rows with low coverage. Then we automatically filter a few rows with the lowest coverage and repeat the attempt. We do so until an attempt of the gradient descend succeeds.
}
