% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dextbetabinom_max_likelihood_Q_attempt.R
\name{dextbetabinom_max_likelihood_Q_attempt}
\alias{dextbetabinom_max_likelihood_Q_attempt}
\title{Attempt To Find Q Maximizing The Likelihood}
\usage{
dextbetabinom_max_likelihood_Q_attempt(
  df,
  starting_Q,
  inf_Q,
  sup_Q,
  derivative_to_diff_const,
  streak_base,
  oob_tax,
  max_step,
  final_accuracy
)
}
\arguments{
\item{df}{A dataframe with the maternal and paternal counts and allelic imbalance proportions.}

\item{starting_Q}{Starting Q for the graident descent.}

\item{inf_Q}{The lowest Q allowed for the gradient descent.}

\item{sup_Q}{The highest Q allowed for the gradient descent.}

\item{derivative_to_diff_const}{The constant for transitioning from the derivative to the step of the gradient descend.}

\item{streak_base}{The exponent base speeding up gradient descent going consecutively in one direction.}

\item{oob_tax}{The tax on any step going out of the predefined bounds.}

\item{max_step}{The bound on the amount of steps in the gradient descent.}

\item{final_accuracy}{The target accuracy to finish the gradient step.}
}
\value{
The list with the final Q of the gradient descent and the success status.
}
\description{
Finds a maximizing Q using one attempt at gradient descend.
}
