% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dextbetabinom_positive_range_Q.R
\name{dextbetabinom_positive_range_Q}
\alias{dextbetabinom_positive_range_Q}
\title{Positivity Range of Extended Beta-Binomial Distribution in Terms of Q}
\usage{
dextbetabinom_positive_range_Q(x, size, AI)
}
\arguments{
\item{x}{The number of successful draws.}

\item{size}{The total number of draws.}

\item{AI}{The proportion between alpha and beta.}
}
\value{
A tuple representing the infimum and the supremum of all values of Q where the pdf is positive.
}
\description{
Returns the interval of Q where the probability distribution function of the extended beta-binomial distribution is positive. Useful later for the gradient descend procedures.
}
